(chore): 시각적 성취도 common 빌드 수정
